fix: resolve all golangci-lint errors (unconvert, gofmt, unparam)
- ftps.go: remove int64(size) casts — FileSize() already returns int64
- blob.go, s3.go, applog.go: run gofmt to fix alignment-space formatting
  (extra spaces used to visually align const values are not gofmt-style)
- flags.go: change applyCliFlags signature from (cfg) error to (cfg)
  since the function never returns a non-nil error (unparam); move all
  validation responsibility to validateConfig where it belongs
- flags_test.go, transfer.go: remove now-dead error handling at all
  applyCliFlags call sites

+++ internal/infrastructure/storage/ftps.go
@@ -107,7 +107,7 @@ func (f *FTPSStorage) Read(ctx context.Context, path string) (io.ReadCloser, int
 		return nil, 0, fmt.Errorf("failed to retrieve file: %w", err)
 	}
 
-	return resp, int64(size), nil
+	return resp, size, nil
 }
 
 func (f *FTPSStorage) Stat(ctx context.Context, path string) (*entity.FileInfo, error) {
@@ -128,7 +128,7 @@ func (f *FTPSStorage) Stat(ctx context.Context, path string) (*entity.FileInfo, 
 	return &entity.FileInfo{
 		Path:         path,
 		Name:         filepath.Base(path),
-		Size:         int64(size),
+		Size:         size,
 		ModifiedTime: modTime,
 		IsDirectory:  false,
 	}, nil
